requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.13.1
Flask==3.1.0
fonttools==4.56.0
fsspec==2024.6.1
grad-cam==1.5.4
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.2
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
requests==2.32.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.3.13
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchmetrics==1.7.0
torchvision==0.21.0+cu124
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3

